requirements.txt
addict==2.4.0
antialiased-cnns==0.3
autopep8==1.6.0
backcall==0.2.0
black==21.9b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.0
efficientnet-pytorch==0.7.1
entrypoints==0.3
h5py==3.5.0
idna==2.10
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numpy==1.21.0
opencv-python==4.5.2.54
packaging==21.2
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.4.0
prompt-toolkit==3.0.21
pycocotools==2.0.2
pycodestyle==2.8.0
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
pywin32==301
PyYAML==6.0
pyzmq==22.3.0
regex==2021.10.23
requests==2.25.1
scipy==1.7.2
shell==1.0.1
six==1.16.0
toml==0.10.2
tomli==1.2.2
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
yapf==0.31.0
